frenchTarot: code restructuring
