+++ web/src/games/frenchTarot/types.ts
