+++ web/src/games/frenchTarot/util/placement.ts
@@ -1,5 +1,5 @@
-import { IG, CardColor } from './types';
-import * as util from './util';
+import { IG, CardColor } from '../types';
+import * as util from './misc';
 
 export function selectableCards(G: IG, playerId: string): boolean[] {
   const player = util.getPlayerById(G, playerId);
