+++ web/src/games/frenchTarot/util/poignee.ts
@@ -1,7 +1,7 @@
 import { Ctx } from 'boardgame.io';
 
-import { IG, ICard, CardColor } from './types';
-import * as util from './util';
+import { IG, ICard, CardColor } from '../types';
+import * as util from './misc';
 
 export function selectableCards(G: IG, playerId: string): boolean[] {
   const player = util.getPlayerById(G, playerId);
